server/src/requirements.txt
pyramid
pyramid_jinja2
mysql-connector-python
python-dotenv